Move the accesstoken data class to its own pkg
