+++ library/src/main/java/de/rheinfabrik/heimdall2/rxjava/OAuth2AccessTokenManager.kt
@@ -1,5 +1,6 @@
 package de.rheinfabrik.heimdall2.rxjava
 
+import de.rheinfabrik.heimdall2.accesstoken.OAuth2AccessToken
 import de.rheinfabrik.heimdall2.rxjava.grants.OAuth2Grant
 import de.rheinfabrik.heimdall2.rxjava.grants.OAuth2RefreshAccessTokenGrant
 import io.reactivex.Single
