+++ library/src/main/java/de/rheinfabrik/heimdall2/rxjava/OAuth2AccessTokenStorage.kt
@@ -1,5 +1,6 @@
 package de.rheinfabrik.heimdall2.rxjava
 
+import de.rheinfabrik.heimdall2.accesstoken.OAuth2AccessToken
 import io.reactivex.Single
 
 /**
