+++ library/src/test/java/de/rheinfabrik/heimdall2/rxjava/OAuth2AccessTokenIsExpiredTest.kt
@@ -1,6 +1,6 @@
 package de.rheinfabrik.heimdall2.rxjava
 
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessToken
+import de.rheinfabrik.heimdall2.accesstoken.OAuth2AccessToken
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.util.*
