+++ library/src/test/java/de/rheinfabrik/heimdall2/rxjava/OAuth2AccessTokenManagerGetValidAccessTokenTest.kt
@@ -3,6 +3,7 @@ package de.rheinfabrik.heimdall2.rxjava
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
+import de.rheinfabrik.heimdall2.accesstoken.OAuth2AccessToken
 import de.rheinfabrik.heimdall2.rxjava.grants.OAuth2RefreshAccessTokenGrant
 import io.reactivex.Single
 import org.junit.Test
