+++ sample/src/main/java/de/rheinfabrik/heimdalldroid/network/oauth2/TraktTvAuthorizationCodeGrant.java
@@ -6,7 +6,7 @@ import de.rheinfabrik.heimdall2.rxjava.grants.OAuth2AuthorizationCodeGrant;
 import io.reactivex.Observable;
 import java.net.URL;
 
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessToken;
+import de.rheinfabrik.heimdall2.accesstoken.OAuth2AccessToken;
 import de.rheinfabrik.heimdalldroid.network.TraktTvApiFactory;
 import de.rheinfabrik.heimdalldroid.network.models.AccessTokenRequestBody;
