+++ sample/src/main/java/de/rheinfabrik/heimdalldroid/network/oauth2/TraktTvOauth2AccessTokenManager.java
@@ -3,7 +3,7 @@ package de.rheinfabrik.heimdalldroid.network.oauth2;
 import android.content.Context;
 import android.content.SharedPreferences;
 
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessToken;
+import de.rheinfabrik.heimdall2.accesstoken.OAuth2AccessToken;
 import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessTokenManager;
 import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessTokenStorage;
 import de.rheinfabrik.heimdalldroid.TraktTvAPIConfiguration;
